+++ climpred/comparisons.py
@@ -31,39 +31,61 @@ def _drop_members(ds, rmd_member=None):
     return ds.sel(member=member_list)
 
 
-class Comparison:
-    'Master class for all comparisons.'
+def _display_comparison_metadata(self):
+    summary = '----- Comparison metadata -----\n'
+    summary += f'Name: {self.name}\n'
+    # probabilistic or only deterministic
+    if not self.probabilistic:
+        summary += 'Kind: deterministic\n'
+    else:
+        summary += 'Kind: deterministic and probabilistic\n'
+    summary += f'long_name: {self.long_name}\n'
+    # doc
+    summary += f'Function: {self.function.__doc__}\n'
+    return summary
 
-    def __init__(self, name, function, is_hindcast, is_probabilistic, longname=None):
+
+class Comparison:
+    """Master class for all comparisons."""
+
+    def __init__(self, name, function, hindcast, probabilistic, long_name=None):
         """Comparison initialization.
 
         Args:
             name (str): name of comparison.
             function (function): comparison function.
-            is_hindcast (bool): Can comparison be used in `compute_hindcast`?
+            hindcast (bool): Can comparison be used in `compute_hindcast`?
              `False` means `compute_perfect_model`
-            is_probabilistic (bool): Is comparison probabilistic? `False` means
-             deterministic.
-            longname (str, optional): Longname of comparison. Defaults to None.
+            probabilistic (bool): Can this comparison be used for probabilistic
+             metrics also? Probabilistic metrics require multiple forecasts.
+             `False` means that comparison is only deterministic.
+             `True` means that comparison can be used both deterministic and
+             probabilistic.
+            long_name (str, optional): longname of comparison. Defaults to None.
 
         Returns:
             comparison: comparison class Comparison.
 
         """
         self.name = name
-        self.longname = longname
         self.function = function
-        self.is_probabilistic = is_probabilistic
-        self.is_hindcast = is_hindcast
+        self.hindcast = hindcast
+        self.probabilistic = probabilistic
+        self.long_name = long_name
+
+    def __repr__(self):
+        """Show metadata of comparison class."""
+        return _display_comparison_metadata(self)
 
 
 # --------------------------------------------#
 # PERFECT-MODEL COMPARISONS
-# based on supervector approach
 # --------------------------------------------#
+
+
 def _m2m(ds, stack_dims=True):
     """
-    Create two supervectors to compare all members to all others in turn.
+    Compare all members to all others in turn while leaving out the verification member.
 
     Args:
         ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
@@ -102,20 +124,19 @@ def _m2m(ds, stack_dims=True):
     return forecast, reference
 
 
-_m2m = Comparison(
+__m2m = Comparison(
     name='m2m',
-    longname='Comparison of multiple forecasts vs. multiple verifications',
     function=_m2m,
-    is_hindcast=False,
-    is_probabilistic=True,
+    hindcast=False,
+    probabilistic=True,
+    long_name='Comparison of multiple forecasts vs. multiple verifications',
 )
 
 
 def _m2e(ds, stack_dims=True):
-    # stack_dims
     """
-    Create two supervectors to compare all members to ensemble mean while
-     leaving out the reference when creating the forecasts.
+    Compare all members to ensemble mean while leaving out the reference in
+     ensemble mean.
 
     Args:
         ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
@@ -142,18 +163,18 @@ def _m2e(ds, stack_dims=True):
     return forecast, reference
 
 
-_m2e = Comparison(
+__m2e = Comparison(
     name='m2e',
-    longname='Comparison of the ensemble mean forecast vs. multiple verifications',
     function=_m2e,
-    is_hindcast=False,
-    is_probabilistic=False,
+    hindcast=False,
+    probabilistic=False,
+    long_name='Comparison of multiple verifications vs. the ensemble mean forecast',
 )
 
 
 def _m2c(ds, control_member=None, stack_dims=True):
     """
-    Create two supervectors to compare all members to control.
+    Compare all other members to control member.
 
     Args:
         ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
@@ -177,18 +198,18 @@ def _m2c(ds, control_member=None, stack_dims=True):
     return forecast, reference
 
 
-_m2c = Comparison(
+__m2c = Comparison(
     name='m2c',
-    longname='Comparison of control forecast vs. multiple verifications',
     function=_m2c,
-    is_hindcast=False,
-    is_probabilistic=True,
+    hindcast=False,
+    probabilistic=True,
+    long_name='Comparison of multiple forecasts vs. control verification',
 )
 
 
 def _e2c(ds, control_member=None, stack_dims=True):
     """
-    Create two supervectors to compare ensemble mean to control.
+    Compare ensemble mean to control member.
 
     Args:
         ds (xarray object): xr.Dataset/xr.DataArray with member and ensemble
@@ -213,18 +234,17 @@ def _e2c(ds, control_member=None, stack_dims=True):
     return forecast, reference
 
 
-_e2c = Comparison(
+__e2c = Comparison(
     name='e2c',
-    longname='Comparison of the ensemble mean forecast vs. control as verification',
     function=_e2c,
-    is_hindcast=False,
-    is_probabilistic=False,
+    hindcast=False,
+    probabilistic=False,
+    long_name='Comparison of the ensemble mean forecast vs. control as verification',
 )
 
 
 # --------------------------------------------#
-# REFERENCE COMPARISONS
-# based on supervector approach
+# HINDCAST COMPARISONS
 # --------------------------------------------#
 def _e2r(ds, reference, stack_dims=True):
     """
@@ -248,12 +268,12 @@ def _e2r(ds, reference, stack_dims=True):
     return forecast, reference
 
 
-_e2r = Comparison(
+__e2r = Comparison(
     name='e2r',
-    longname='Comparison of the ensemble mean vs. reference verification',
     function=_e2r,
-    is_hindcast=True,
-    is_probabilistic=False,
+    hindcast=True,
+    probabilistic=False,
+    long_name='Comparison of the ensemble mean vs. reference verification',
 )
 
 
@@ -284,13 +304,13 @@ def _m2r(ds, reference, stack_dims=True):
     return forecast, reference
 
 
-_m2r = Comparison(
+__m2r = Comparison(
     name='m2r',
-    longname='Comparison of multiple forecasts vs. reference verification',
     function=_m2r,
-    is_hindcast=True,
-    is_probabilistic=True,
+    hindcast=True,
+    probabilistic=True,
+    long_name='Comparison of multiple forecasts vs. reference verification',
 )
 
 
-__all_comparisons__ = [_m2m, _m2e, _m2c, _e2c, _e2r, _m2r]
+__ALL_COMPARISONS__ = [__m2m, __m2e, __m2c, __e2c, __e2r, __m2r]