+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ Depreciated
 
 Features
 --------
-- re-written ``metrics`` and ``comparisons`` based on classes to allow user metrics and user comparisons. (:pr:`2xx`) `Aaron Spring`_.
+- Re-wrote ``metrics`` and ``comparisons`` based on classes to allow user metrics and user comparisons. See https://climpred.readthedocs.io/en/stable/metrics.html##user-defined-metrics and https://climpred.readthedocs.io/en/stable/comparisons.html##user-defined-comparisons. (:pr:`268`) `Aaron Spring`_.
 - `metric_kwargs` are passed to `xs.metric`. (:pr:`264`) `Aaron Spring`_.
 - New deterministic metrics: `spearman_r`, `spearman_r_p_value`, `mad`, `mape`, `smape`. (:pr:`264`) `Aaron Spring`_.
 - Rewrote `varweighted_mean_period` based on `xrft`. Changed `time_dim` to `dim` in `stats.varweighted_mean_period`. Keeps coords. (:pr:`258`) `Aaron Spring`_.